app_build/pyproject.toml
[project]
name            = "gimp-custom"
version         = "0.1.0"
description     = "A collection of GIMP plugins that make use of generative ai models, image editing, and more"
readme          = "README.md"
requires-python = ">=3.12"
dependencies    = [
    "google-genai>=2.0.0",
    "pygobject>=3.56.2",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = ["pytest>=7.4.0", "requests>=2.31.0", "pygobject-stubs>=2.17.0"]
